Updates
I've made some modifications all over the code to make it easyer to be readable and used.
I've also figured why pin PA15, PB3 and PB4 were not working. By default the MCU configure this pins to be used with a JTAG, but as I'm using only St-link and i'll need these pins as PWM in my project, I've putted a default command at gpio::configpin to as soon as this function is called, it enable this pins to be used by other periferals. Only PA13 and PA14 still being used by st-link